rudimentary home page update
home page edit subpage is now routed through a subgroup. added a route for updating home page information

added a controller to oversee home page operations that provides methods for both routes

added a functional form to the home page edit subpage. note that the formatting is basic at this time and will need to be improved. it only contains four of the many fields that it will eventually include