added domain and subdomains
for testing purposes the code now assumes this project is hosted at the domain ftb.sites

all main app routes are now mapped to the websitebuilder subdomain

added a unique subdomain field to the websites table. this included updating the migration, updating the database seeder, and updating the model

the generated sites are now hosted at [subdomain].ftb.sites (where [subdomain] is determined by the new subdomain field). added routes and controller methods to facilitate this, as well as editing the preview components and controller methods

note that for this to work locally the etc/hosts file has to include mappings for each subdomain (wildcard domains are not supported locally but should be fine in an actual DNS record)